requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv 
httpx 
pydantic-settings
pydantic[email]
python-multipart